Working or KYC and Payment module

+++ src/main/java/com/appbackend/example/AppBackend/entities/UserLoanEligibility.java
@@ -17,7 +17,7 @@ public class UserLoanEligibility {
     private LoanEligibility eligibility;
 
     @Column(name="eligibility_amount")
-    private Double eligibilityAmount;
+    private Long eligibilityAmount;
 
     @OneToOne
     private User user;
